F04/multiple projects (#9)
* Changed the chache to store each Project in seperate keys

* Re-uploading the same Project now overwrites the old one

* Setup Firebase test

* Final Cache structure

* EXTENDED CHITIN [2/3EP]
The operative gains the following ability for the battle:

Extended Chitin: Each time a shooting attack is made against this operative, in the Roll Defence Dice step of that shooting attack, you can re-roll one of your defence dice.

If the operative is a TYRANID WARRIOR, this equipment costs 3EP; otherwise, it costs 2EP.

* support for multiple projects + minor refactoring

* Functionality yes, looks good no

-Tried out themes, but only cost me 2 hours of my lifetime because cannot un-scuff widget

* Quick cleanup of useless node files
