+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "firebase": "^9.6.7"
+  }
+}
